up swap deployment
